pyInterface: rename internal Python files
Rename the internal Python files and put an underscore in front of
their current name. The underscore should indicate that not the whole
file is to be imported, but only the selected functions.
