+++ pyInterface/package/__init__.py
@@ -1,20 +1,21 @@
 
 import ROOT
 
+from _amplitude import calcAmplitude
 from _config import rootPwaConfig
 from _fileManager import fileManager
 from _fileManager import saveFileManager
 from _fileManager import loadFileManager
-from amplitude import calcAmplitude
-from fit import pwaFit
-from fit import pwaNloptFit
-from integrals import calcIntegrals
-from likelihood import initLikelihood
+from _fit import pwaFit
+from _fit import pwaNloptFit
+from _integrals import calcIntegrals
+from _likelihood import initLikelihood
 
+del _amplitude
 del _config
 del _fileManager
-del amplitude
-del integrals
-del fit
+del _fit
+del _integrals
+del _likelihood
 
 config = None
