+++ pyInterface/package/_amplitude.py
